+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Remove version suffix from reference version before updating `project.go`.
+
 ## [2.1.3] - 2020-08-11
 
 ### Fixed